+++ pullbug/messages.py
@@ -1,36 +1,40 @@
-"""Pull Bug Sending Messages Logic"""
 import os
-import sys
+import logging
 import requests
-from dotenv import load_dotenv
 import slack
 
 
-class Messages():
-    """All sending message logic lives here"""
-    # Setup variables
-    load_dotenv()
-    SLACK_BOT_TOKEN = os.getenv('SLACK_BOT_TOKEN')
-    ROCKET_CHAT_URL = os.getenv('ROCKET_CHAT_URL')
+SLACK_BOT_TOKEN = os.getenv('SLACK_BOT_TOKEN')
+SLACK_CHANNEL = os.getenv('SLACK_CHANNEL')
+ROCKET_CHAT_URL = os.getenv('ROCKET_CHAT_URL')
+LOGGER = logging.getLogger(__name__)
 
+
+class Messages():
     @classmethod
-    def rocket_chat(cls, message):
-        """Send a Rocket Chat message"""
+    def rocketchat(cls, message):
+        """Send a Rocket Chat message
+        """
         try:
-            requests.post(Messages.ROCKET_CHAT_URL, data={'text': message})
-            print("Rocket Chat message sent!")
+            requests.post(ROCKET_CHAT_URL, data={'text': message})
+            LOGGER.info('Rocket Chat message sent!')
         except requests.exceptions.RequestException as rc_error:
-            sys.exit(rc_error)
+            LOGGER.warning(f'Could not send Rocket Chat message: {rc_error}')
+            raise requests.exceptions.RequestException(rc_error)
 
     @classmethod
     def slack(cls, message):
-        """Send Slack messages via a bot"""
-        slack_client = slack.WebClient(Messages.SLACK_BOT_TOKEN)
+        """Send Slack messages via a bot
+        """
+        slack_client = slack.WebClient(SLACK_BOT_TOKEN)
         try:
             slack_client.chat_postMessage(
-                channel=os.getenv("SLACK_CHANNEL"),
+                channel=SLACK_CHANNEL,
                 text=message
             )
-            print("Slack message sent!")
+            LOGGER.info('Slack message sent!')
         except slack.errors.SlackApiError as slack_error:
-            sys.exit(slack_error)
+            LOGGER.warning(f'Could not send Slack message: {slack_error}')
+            raise slack.errors.SlackApiError(
+                slack_error.response["ok"], slack_error.response['error']
+            )